requirements.txt
appnope==0.1.2
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
certifi==2021.5.30
charset-normalizer==2.0.6
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.0
Django==3.2.8
django-debug-toolbar==3.2.2
django-environ==0.7.0
django-split-settings==1.1.0
djangorestframework==3.12.4
drf-yasg==1.20.0
eradicate==2.0.0
Faker==9.2.0
flake8==3.9.2
flake8-eradicate==1.1.0
flake8-print==4.0.0
idna==3.2
inflection==0.5.1
iniconfig==1.1.1
ipython==7.28.0
isort==5.9.3
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.3
loguru==0.5.3
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mixer==7.1.2
packaging==21.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.20
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
traitlets==5.1.0
uritemplate==4.1.1
urllib3==1.26.7
wcwidth==0.2.5
